move abstract classes to base package
