+++ data/src/test/java/openweather/data/base/BaseMockTest.kt
@@ -1,4 +1,4 @@
-package openweather.data
+package openweather.data.base
 
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
